Move internal types into Google.Cloud.Spanner.V1.Internal
