+++ apis/Google.Cloud.Spanner.Data/Google.Cloud.Spanner.Data.IntegrationTests/StressTestBase.cs
@@ -19,7 +19,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 using Google.Cloud.Spanner.V1;
-using Google.Cloud.Spanner.V1.Logging;
+using Google.Cloud.Spanner.V1.Internal.Logging;
 using Xunit;
 
 namespace Google.Cloud.Spanner.Data.IntegrationTests
