+++ apis/Google.Cloud.Spanner.Data/Google.Cloud.Spanner.Data.IntegrationTests/TestLogger.cs
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 using System;
-using Google.Cloud.Spanner.V1.Logging;
+using Google.Cloud.Spanner.V1.Internal.Logging;
 using Xunit.Abstractions;
 
 namespace Google.Cloud.Spanner.Data.IntegrationTests
