+++ apis/Google.Cloud.Spanner.Data/Google.Cloud.Spanner.Data/ExecuteHelper.cs
@@ -16,7 +16,7 @@ using System;
 using System.Diagnostics;
 using System.Threading;
 using System.Threading.Tasks;
-using Google.Cloud.Spanner.V1.Logging;
+using Google.Cloud.Spanner.V1.Internal.Logging;
 
 namespace Google.Cloud.Spanner.Data
 {
