+++ apis/Google.Cloud.Spanner.Data/Google.Cloud.Spanner.Data/SpannerException.cs
@@ -15,8 +15,8 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
-using Google.Cloud.Spanner.V1;
-using Google.Cloud.Spanner.V1.Logging;
+using Google.Cloud.Spanner.V1.Internal;
+using Google.Cloud.Spanner.V1.Internal.Logging;
 using Grpc.Core;
 
 namespace Google.Cloud.Spanner.Data
