+++ apis/Google.Cloud.Spanner.V1/Google.Cloud.Spanner.V1/Internal/Logging/LogLevel.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace Google.Cloud.Spanner.V1.Logging
+namespace Google.Cloud.Spanner.V1.Internal.Logging
 {
     /// <summary>
     /// Log levels for Spanner client internal logging.
